Interface graphique pion 2

+++ src/Configuration.java
@@ -2,8 +2,8 @@
 public class Configuration {
     public static int MAXMAPSIZE;
     public static int MAPSIZE;
-    public static double WIDTH;
-    public static double HEIGHT;
+    public static int WIDTH;
+    public static int HEIGHT;
     public static int PIONSIZE=20;
     public static boolean END=false;
 }
